requirements.txt
pillow
pytesseract
stegano
exifread
python-magic
python-magic-bin; platform_system=="Windows"
mutagen
tkinterdnd2
